wip - finding integration test bug
