+++ integration_tests/mockApis/prisonApi.ts
@@ -2,7 +2,7 @@ import { SuperAgentRequest } from 'superagent'
 import { stubFor } from './wiremock'
 import { CaseLoad } from '../../server/data/prisonApiClient'
 import { Agency } from '../../server/data/PrisonLocationResult'
-import { alertCodeString } from '../../server/utils/alertHelper'
+// import { alertCodeString } from '../../server/utils/alertHelper'
 
 const stubPing = (status = 200): SuperAgentRequest =>
   stubFor({
@@ -115,19 +115,18 @@ const stubGetUsersLocations = (response = []): SuperAgentRequest =>
     },
   })
 
-const stubGetPrisonersAlerts = ({ prisonerNumber, response = [], status = 200 }): SuperAgentRequest =>
-  stubFor({
-    request: {
-      method: 'GET',
-      // url: `/prisonApi/api/offenders/${prisonerNumber}/alerts/v2?alertCodes=${alertCodeString}`,
-      url: `/offenders/${prisonerNumber}/alerts?alertCode=${alertCodeString}`,
-    },
-    response: {
-      status,
-      headers: { 'Content-Type': 'application/json;charset=UTF-8' },
-      jsonBody: response,
-    },
-  })
+// const stubGetPrisonersAlerts = ({ prisonerNumber, response = [], status = 200 }): SuperAgentRequest =>
+//   stubFor({
+//     request: {
+//       method: 'GET',
+//       url: `/prisonApi/api/offenders/${prisonerNumber}/alerts/v2?alertCodes=${alertCodeString}`,
+//     },
+//     response: {
+//       status,
+//       headers: { 'Content-Type': 'application/json;charset=UTF-8' },
+//       jsonBody: response,
+//     },
+//   })
 
 const stubGetMovementByOffender = ({ response = [], status = 200 }): SuperAgentRequest =>
   stubFor({
@@ -169,7 +168,7 @@ export default {
   stubGetSecondaryLanguages,
   stubGetBatchPrisonerDetails,
   stubGetUsersLocations,
-  stubGetPrisonersAlerts,
+  // stubGetPrisonersAlerts,
   stubGetMovementByOffender,
   stubValidateChargeNumber,
 }
